requirements.txt
gym >= 0.25.0
pygame >= 2.1.2
h5py >= 1.21.0
hydra-core >= 1.2.0
tqdm >= 4.64.0
mujoco >= 2.2.0
scipy >= 1.8.0
tensorboard >= 2.9.1